problema de casting em MenuProdutos

+++ byte_brew/src/br/edu/cafeteria/app/MenuProdutos.java
@@ -592,7 +592,7 @@ public class MenuProdutos {
     		   				break;
     		   				
     		   			}
-    		   			tempo = Integer.parseInt(entrada);
+    		   			tempo = Double.parseDouble(entrada);
     		   			comida.setTempoPreparo(tempo);
     		   			
     		   			if(cadastroProduto.editarProduto(comida)) {
